+++ Formula/netductor.rb
@@ -1,26 +1,26 @@
 class Netductor < Formula
   desc "Netductor control plane CLI / TUI"
   homepage "https://github.com/PavelNeyman/netductor"
-  version "0.8.79"
+  version "0.8.80"
   license "MIT"
   on_macos do
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.79/netductor-darwin-arm64"
-      sha256 "e2547bb0b02d189f11d0f132f6cadc7774e1534285f1ef8ba92cee1c61ad42dc"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.80/netductor-darwin-arm64"
+      sha256 "5fdc1f8a7e25649a75a59ddb18f771930aa1502b568dc98cd92995df25dc9c7e"
     end
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.79/netductor-darwin-amd64"
-      sha256 "45440667a20d4238f3e3814a776139a476f6f7f693c1f9a6f5b32bc7b57e3d0b"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.80/netductor-darwin-amd64"
+      sha256 "04bc16dce918cc1de50f0eb9681e3c44d63ec47c88e6a845e9420ab245095338"
     end
   end
   on_linux do
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.79/netductor-linux-amd64"
-      sha256 "9035b73e571362d2667bf1e777e28f3dbecdc38319a8ffb87953e5b04e122042"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.80/netductor-linux-amd64"
+      sha256 "16f05c109602acb217f707b31398b3c921c1b3bd699ed9985d39f59021cc4011"
     end
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.79/netductor-linux-arm64"
-      sha256 "c05acf985e5ecb907f7df84c66380d83afb4ed03707a4622134793bb0e4f29d0"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.80/netductor-linux-arm64"
+      sha256 "1cd1e90b75fb7d312c9d1bb64c8e039764ba862aace04f321a55ab06cc72a463"
     end
   end
   def install